# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PlanDayOut": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/PlanEntryOut"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "date",
          "entries"
        ],
        "title": "PlanDayOut",
        "type": "object"
      },
      "PlanDaysOut": {
        "description": "GET /api/plan/days?month=YYYY-MM: days with meals, for the calendar dots.",
        "properties": {
          "dates": {
            "items": {
              "format": "date",
              "type": "string"
            },
            "title": "Dates",
            "type": "array"
          },
          "month": {
            "title": "Month",
            "type": "string"
          }
        },
        "required": [
          "month",
          "dates"
        ],
        "title": "PlanDaysOut",
        "type": "object"
      },
      "PlanEntryCreate": {
        "description": "POST /api/plan/entries. The meal goes last on that day.",
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "recipeId": {
            "format": "uuid",
            "title": "Recipeid",
            "type": "string"
          },
          "servings": {
            "minimum": 1.0,
            "title": "Servings",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "recipeId",
          "servings"
        ],
        "title": "PlanEntryCreate",
        "type": "object"
      },
      "PlanEntryOut": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "recipeDeleted": {
            "title": "Recipedeleted",
            "type": "boolean"
          },
          "recipeId": {
            "format": "uuid",
            "title": "Recipeid",
            "type": "string"
          },
          "recipeName": {
            "title": "Recipename",
            "type": "string"
          },
          "recipeSlug": {
            "title": "Recipeslug",
            "type": "string"
          },
          "servings": {
            "title": "Servings",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "date",
          "position",
          "recipeId",
          "recipeSlug",
          "recipeName",
          "recipeDeleted",
          "servings"
        ],
        "title": "PlanEntryOut",
        "type": "object"
      },
      "PlanEntryUpdate": {
        "description": "PATCH /api/plan/entries/{id}: change servings, swap recipe, move day or reorder.",
        "properties": {
          "date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "position": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "recipeId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipeid"
          },
          "servings": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servings"
          }
        },
        "title": "PlanEntryUpdate",
        "type": "object"
      },
...
      "WeekPlanOut": {
        "description": "GET /api/plan?week=YYYY-MM-DD (any date in the week; defaults to this week).",
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/PlanDayOut"
            },
            "title": "Days",
            "type": "array"
          },
          "weekStart": {
            "format": "date",
            "title": "Weekstart",
            "type": "string"
          }
        },
        "required": [
          "weekStart",
          "days"
        ],
        "title": "WeekPlanOut",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Meal planning for households on a tight budget. Errors always use the `ErrorResponse` object.",
    "title": "Nosh API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/plan": {
      "get": {
        "description": "Monday to Sunday with each day's meals in order. `week` can be any date in the week; defaults to today.",
        "operationId": "get_week",
        "parameters": [
          {
            "description": "Any date in the week (YYYY-MM-DD)",
            "in": "query",
            "name": "week",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Any date in the week (YYYY-MM-DD)",
              "title": "Week"
            }
          },
          {
            "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
            "in": "query",
            "name": "today",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
              "title": "Today"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekPlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Get a week's plan",
        "tags": [
          "Week plan"
        ]
      }
    },
    "/api/plan/days": {
      "get": {
        "description": "For the calendar picker's dots.",
        "operationId": "planned_days",
        "parameters": [
          {
            "description": "YYYY-MM",
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "description": "YYYY-MM",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "title": "Month",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDaysOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Days with meals in a month",
        "tags": [
          "Week plan"
        ]
      }
    },
    "/api/plan/entries": {
      "post": {
        "description": "The meal goes last on that day. Unlimited meals per day.",
        "operationId": "add_entry",
        "parameters": [
          {
            "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
            "in": "query",
            "name": "today",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
              "title": "Today"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanEntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Add a meal to a day",
        "tags": [
          "Week plan"
        ]
      }
    },
    "/api/plan/entries/{entry_id}": {
      "delete": {
        "operationId": "delete_entry",
        "parameters": [
          {
            "description": "Plan entry id",
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "description": "Plan entry id",
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          },
          {
            "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
            "in": "query",
            "name": "today",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
              "title": "Today"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Remove a meal from the plan",
        "tags": [
          "Week plan"
        ]
      },
      "patch": {
        "description": "Change servings, swap the recipe, move it to another day (goes last there) or reorder it within its day.",
        "operationId": "update_entry",
        "parameters": [
          {
            "description": "Plan entry id",
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "description": "Plan entry id",
              "minimum": 1,
              "title": "Entry Id",
              "type": "integer"
            }
          },
          {
            "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
            "in": "query",
            "name": "today",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The user's local date (YYYY-MM-DD). Defaults to the server's local date.",
              "title": "Today"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanEntryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanEntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error: `details.fields` maps each field to a message"
          }
        },
        "summary": "Change a planned meal",
        "tags": [
          "Week plan"
        ]
      }
    },
...
  },
...
}